Consolidate preview scripts
