+++ .github/ts-workflows/workflows/deploy-streams-example-app.ts
@@ -1,5 +1,5 @@
 import type { Workflow } from "@jlarky/gha-ts/workflow-types";
-import { cloudflarePreviewApps } from "../../../scripts/preview/apps.ts";
+import { cloudflarePreviewApps } from "../../../scripts/preview/preview.ts";
 import { createCloudflareAppWorkflow } from "../utils/cloudflare-app-workflow.ts";
 
 const workflow: Workflow = await createCloudflareAppWorkflow(
