--- scripts/preview/reconcile-environment-config-leases.test.ts
@@ -1,152 +1,0 @@
-import { describe, expect, it } from "vitest";
-import {
-  evaluateCloudflareZoneCheck,
-  reconcileEnvironmentConfigLeaseResources,
-} from "./reconcile-environment-config-leases.ts";
-import { ENVIRONMENT_CONFIG_LEASE_RESOURCE_TYPE } from "./preview-inventory.ts";
-
-describe("reconcileEnvironmentConfigLeaseResources", () => {
-  it("checks live Semaphore leases against Doppler projects and Cloudflare zones", async () => {
-    const result = await reconcileEnvironmentConfigLeaseResources({
-      client: {
-        list: async () => [
-          {
-            slug: "preview-2",
-            data: { dopplerConfig: "preview_2" },
-            leaseState: "available",
-            leasedUntil: null,
-          },
-        ],
-      },
-      checkDopplerConfig: async () => ({ ok: true }),
-      readCloudflareCredentials: async ({ project }) => ({
-        ok: true,
-        project,
-        accountId: "cf-account",
-        apiToken: "redacted",
-      }),
-      checkCloudflareZone: async () => ({ ok: true }),
-      commandEnvironment: {},
-      repositoryRoot: "/repo",
-      semaphoreBaseUrl: "https://semaphore.iterate.com",
-    });
-
-    expect(result).toMatchObject({
-      ok: true,
-      type: ENVIRONMENT_CONFIG_LEASE_RESOURCE_TYPE,
-      summary: {
-        issueCount: 0,
-        resourceCount: 1,
-      },
-      resources: [
-        {
-          slug: "preview-2",
-          dopplerConfig: "preview_2",
-          domains: ["iterate-preview-2.com", "iterate-preview-2.app"],
-          issues: [],
-        },
-      ],
-    });
-  });
-
-  it("reports malformed resource data, missing Doppler configs, and inaccessible zones", async () => {
-    const result = await reconcileEnvironmentConfigLeaseResources({
-      client: {
-        list: async () => [
-          {
-            slug: "preview-2",
-            data: { dopplerConfig: "preview_2", note: "extra data should not live here" },
-            leaseState: "leased",
-            leasedUntil: 1_777_984_800_000,
-          },
-          {
-            slug: "preview-3",
-            data: { dopplerConfig: "preview_3" },
-            leaseState: "available",
-            leasedUntil: null,
-          },
-        ],
-      },
-      checkDopplerConfig: async ({ config, project }) =>
-        project === "os" && config === "preview_3"
-          ? { ok: false, message: "config not found" }
-          : { ok: true },
-      readCloudflareCredentials: async ({ project }) => ({
-        ok: true,
-        project,
-        accountId: "cf-account",
-        apiToken: "redacted",
-      }),
-      checkCloudflareZone: async ({ domain }) =>
-        domain === "iterate-preview-3.app"
-          ? { ok: false, message: "zone not found in Cloudflare account cf-account" }
-          : { ok: true },
-      commandEnvironment: {},
-      repositoryRoot: "/repo",
-      semaphoreBaseUrl: "https://semaphore.iterate.com",
-    });
-
-    expect(result.ok).toBe(false);
-    expect(result.summary.issueCount).toBe(3);
-    expect(result.resources.flatMap((resource) => resource.issues)).toEqual([
-      {
-        check: "resource-data",
-        resourceSlug: "preview-2",
-        message: "Resource data must contain only dopplerConfig.",
-      },
-      {
-        check: "doppler-config",
-        resourceSlug: "preview-3",
-        message: "os/preview_3: config not found",
-      },
-      {
-        check: "cloudflare-zone",
-        resourceSlug: "preview-3",
-        message: "iterate-preview-3.app: zone not found in Cloudflare account cf-account",
-      },
-    ]);
-  });
-});
-
-describe("evaluateCloudflareZoneCheck", () => {
-  it("rejects a moved same-account zone when DNS is delegated to a different active zone", () => {
-    expect(
-      evaluateCloudflareZoneCheck({
-        accountId: "preview-account",
-        domain: "iterate-preview-2.com",
-        zones: [
-          {
-            account: { id: "preview-account" },
-            name: "iterate-preview-2.com",
-            status: "moved",
-          },
-          {
-            account: { id: "delegated-account" },
-            name: "iterate-preview-2.com",
-            status: "active",
-          },
-        ],
-      }),
-    ).toEqual({
-      ok: false,
-      message:
-        "active zone belongs to Cloudflare account delegated-account, expected preview-account",
-    });
-  });
-
-  it("accepts an active zone in the expected account", () => {
-    expect(
-      evaluateCloudflareZoneCheck({
-        accountId: "preview-account",
-        domain: "iterate-preview-2.com",
-        zones: [
-          {
-            account: { id: "preview-account" },
-            name: "iterate-preview-2.com",
-            status: "active",
-          },
-        ],
-      }),
-    ).toEqual({ ok: true });
-  });
-});
